Felix manymouse.
And fix issue with idiot Apple linker spec -framework option
with hack --keepall to flx_pkgconfig. The *right* way to fix this
is actually to put --framework=dir in the *.fpc database, and translate
it to -framework dir when building shell argument lists.
Keepall keeps duplicates, we have to keep two -frameworks.
But we don't want to keep duplicate library specs. Without the
translation we have a conflict. Hopefully we can tolerate
duplicate library specs.