Relax warnings for coercions of compact linear types.
Allow any coercion provided the sizes are the same.
Not really what we want: allows 3 x 4 to be coerced
to 4 x 3 or 6 x 2. We really only want shapely coercions.

In principle, using a coercion is a hack for polyadicity anyhow.
In principle only associators should be allowed, for example:

  (3 * 2) * 2 -> 3 * (2 * 2)

For an array with these indices, in effect this just moves
the "data location" about without otherwise changing the
structure. However, complete flattening is needed for the
polyadic stuff anyhow, and its useful to fill a structured
array from a linear list, etc.