Add a new flag --run-only to flx.
The flag --run-only says to run the executable or dylib,
which would be produced by the command, without any compilation
or dependency checking. It is intended to be used where the program
uses dynamically loaded libraries or plugins and the LD_LIBRARY_PATH
needs to be specified so they can be loaded. flx provides the
appropriate path in a platform independent way, i.e. uses
DYLD_LIBRARY_PATH on OSX.