Add --toolchain= option to flx.
This option names a toolchain plugin to use instead
of the default one obtained from the config database
toolchain.fpc file.

If recompiling with another compiler, you will need to use
the --force switch to defeat the current implementation
of the C++ dependency checking, which sees only include
files and ignores the compiler and its switches.
Unfortunately that will re-run the Felix compiler flxg too.

Note that toolchain plugins are shared libraries which can
be anywhere on your dynamic loader path: in a system default
location or DYLD_LIBRARY_PATH on OSX, LD_LIBRARY_PATH on Linux.