Start to add support for speculative plugin loads.
At present if libraries or symbols are missing,
Felix throws an exception which terminates the program.
Since catching exceptions is not really possible in Felix,
I want to provide functions that return an error code
instead of aborting.

This might be used, for example, by the webserver,
trying to load a formatting routine based on a filename
extension. Plugins can then be written to support some
formats and made available to the webserver whilst it is running.