Bump version. Patch errno handling.
Version 1.1.10 now.

Posix strerror_r() function returns an error code but GNU C library
returns a pointer and so is not posix compliant. To cope with
both, the return type is voided away, which makes it impossible
to tell if the function worked or not.

To fix this problem, the buffer it is supposed to fill in
is now initialised to a failure message reporting the value
of errno. Previously you'd get random junk and possibly
a crash due to a read overrun.