Add LP tool and webserver display.
Add flx_iscr.flx tool and make webserver display
something approximately right for the new commands.

flx_iscr filename optional-directory

processes filename and writes files into optional-directory.
If the latter isn't specified the current directory is used.

Commands have the form:

@tangler nick = filename
@tangle nick
...

The first command defines a tangler with nickname nick,
associated with optional-directory/filename.

The second command populates a buffer associated with the given
nickname with lines following the @tangle command up to the next
@command or the end of file.

Any number of tangle commands  to the
same nick are allowed. The nick must be defined by a tangler command
and tangler commands must only define a nick once, otherwise
the program aborts.

After the input file is processed the aggregated contents of each
tangle buffer are compared to the associated file. If the contents
differ from the file, the buffer is written to the file.

The comparison is done to avoid changing the latest modification
date for the file which may be used in dependency checking.

Any directories are autocreated.

The tool currently cannot distinguish a non-existant file from
a file with empty contents, so if a file doesn't exist and no
data is tangled to it, it will not be created.

The webserver currently recognises only files with ".flx"
or ".cxx" extensions as Felix and C++ files. Any other filename
extension is typeset as preformt data. To be extended later!