Change readln to echo input for file.
The Felix readln function reads cstdin.
If this is connected to a terminal, the input
gets displayed on the terminal, including terminating
newline.

In the old function, however, if you redirect the input
to a file, say for automated testsing, standard output
no longer contains the input. Whilst this is correct,
in a tutorial, display of the "expected output" would
normally have to contain a copy of the input to make
sense to the reader.

The new readln function checks if cstdin's fileno isatty,
and if not, echoes the input to cstdout.

This is implemented in flx_ioutil with three functions:
raw_readln is the old one. A new function echo_readln
checks for a terminal and echos input if not.

Plain old readln just calls echo_readln.

Two functions are used to implement this: flx_fileno
and flx_isatty. They're compiled with conditional
compilation so they work on Windows too. These functions
currently have internal linkage and aren't made available
to avoid an API breakage. However I may export them
later.