Fix parser blank line start fdoc file.
Previously, to parse an fdoc file, Felix required
a blank line at the start. This was because the parser
has no way to detect start of file, and the comment
of the form \n@ ... \n@felix\n requires the first @
in column 1 to avoid parsing a @int nested in a line
as a comment start: @int means a pointer, possibly
NULL, to an int. Can't be used in column 1, but isn't
likely to be either.

The fix uses lexing from a function, which emits a leading
newline. The initial line count is fudged by subtracting 1
to keep everything working.