Change webserver skin.
Minor adjustments, mainly to fdoc. Change colours and sizes
a bit in an attempt to make the docs a bit more attractive.

Remove line numbers from Felix code display because
they get in the way of copying code to the clipboard.
And they weren't sequenced properly anyhow.

Add the tag @likefelix which has the same display
as an @felix tag but will be ignored by the Felix
parser and by the flx_tangle tool. This allows code
fragments to be displayed nicely, without the concatenation
of these fragments in a tutorial having to be a complete
program.