Refactor flx_pkgconfig is preparation for sepcomp.
We need two versions: one which yanks the whole source
in and on which is separately compiled and accessed
through an interface. The latter will be used everywhere
by the Felix build system. The former should only be used
by the Python bootstrap.

At this commit, the separate interface is written but not
used. I need to figure out where to put it as well,
and then write an import module for it. A bit messy!

This is just the first module to split out, so the build
tools can build much faster. At present it seems like half
the time building parts of Felix is taken up first building
the tool used to build it!

Also added a picture to one tutorial. Will play with that
and add some more. It's faster for me, given the crappy
tools I have, to draw diagrams by hand and photograph them
than to try to figure out how to use the diagramming
tools properly. And it looks kinda cute anyhow :)