Fixup shayne's tuts.
Fix the Test lines to reflect the actual tutorial locations.
Fix the boost URL's to refer to the correct tutorials.
Add lots of commentary. Add @expect's when possible.
Clean up the syntax a bit.

Get thinking on how to reference terms to the glossary
or some wiki pages. Also think about how to show
some alternatives easily. For example writing:

proc async_do_after (
    clock:Faio::alarm_clock,
    delta:double,
    p:Faio::alarm_clock->0 )
{
  proc fthread () {
    Faio::sleep (clock, delta);
    p (clock);
  }
  spawn_fthread fthread;
}

is kind of klunky .., we should write instead

  spawn_fthread { Faio::sleep .... };

but that requires knowing what { .. } means.
Whether the reader is expected to know that is unclear.