All tutorials compile OK.
This has two caveats. First, some Felix code is
marked unchecked just so all the tests passed.
They pass by not actually compiling that code!

Second, much of the code doesn't have expect sections
which predict the output. So the code compiler but it
may not work.

The tutorials need to be revisited to fix both these
problems. In particular pthread_01 actually segfaults
for some reason!

I have also added a directory for random articles.
Stuff where I want to talk about things but it doesn't
fit in as reference specification or tutorial material.