Rework flx.
We're officially switching to --felix=file.fpc.
The --test=install should still work at the moment
but will be removed.

BE WARNED the new algorithm applies environment
variables AFTER switches and options. Therefore
DO NOT USE exported global environment variables
to control install directory if you're a developer.
FLX_INSTALL_DIR for example, if set, will clobber
ALL other setting for the library paths and prevent
building. Indeed, that's the purpose: to use a prebuild
script to build something with an alternate library.

I may change this, possibly having TWO sets of environment
variables, one applied early to change the defaults,
and one as an override to influence scripts that can't
be edited.