Try to fix Shayne's when problem.
We add a new function fun andthen(x:bool,y:1->bool)
which is a shortcut and operator taking a closure
for a second argument to ensure it isn't evaluated
prematurely.

Then we rewrite the match checking code to use it.
In the process the special match_check term is
eliminated, and replaced by an ordinary function.

I should do this for the match handler too.