Fdoc browser layout moveable divider.
The vertical boundary between the left column
and right column can now be moved by dragging.

The position is different for every page.
Reloading from the server resets the position.
However at least on Firefox it seems the position
is preserved if the page is reloaded from the cache.

The cosmetics have been changed in the process and
will need a bit of reworking.

The layout changed only applies to fdocs at the moment,
not flx files which use separate code, unfortunately.
So some refactoring is to happen.