Add a routine to detect free fixpoints.
Replace the caching type calculator code
that throws with a non-throwing test.
Add a diagnostic to unfold routine to tell when
a free fixpoint exception is thrown.
Roughly: it should never happen.

During construction of a type, it is inevitable that
a free fixpoint exist on the way up the recursion.
We can't replace it by unfolding because the type is
bound bottom up and we haven't got up to binding the
type up to the implicit fixpoint binder yet.

Note this leads to a very serious problem when an expression
is bound, that a subterm might get a type with an incomplete
fixpoint. In some ways this is inevitable when calculating
function return types. In any case it can't be allowed.