Change graph search to use an abstract ordering.
I used on of those Java like object things,
to create a common interface for a stack and a queue.
Give the algo the queue, you get breadth first.
Give it a stack, depth first.

I may extend it so one can use a priority queue
to optimise shortest path algos.