Add all the set comparisons to Streamable.
This is simply NOT right. But it seems to work.
Somehow, doing this yanks in the value type.
This hack means it won't be possible to implement
subset like operations for non-streambles.

More precisely the LHS of x \subseteq y must be
streamable to support the default implementation
and the RHS need only be a set. So we just run
through all the members of the LHS using the stream
iterator, and check each one is a member of the RHS.

This means, for example, that you cannot do

  RE2 "(a|b)*abb" \subseteq RE2 "(a|b)*"

even if there is a way to do that calculation,
since regexps are sets but not streamable.