Resolve circular dependency.
flx_async.hpp is required by the RTL.
But flx_async requires the RTL.
flx_async is supposed to be an optional add on
for socket I/O and timers etc (because some platforms
can't support this kind of thing).

So I move the header to the RTL but leave the
source in flx_async library.