Split thread control into a base and derived class.
The purpose is to decouple the garbage collector
from the threading system. This is the first step.
Purely for convenience, two things are currenly attached
to the system GC object: the thread control object
and the module registry. It's convenient because the
collector is a singlton: there's only one per process
at the moment.

The thread safe GC and thread control object are
tightly coupled: the GC has to stop threads and
fetch their stacks to operate correctly, as well
as various auxilliary operations being properly serialised.

however the GC doesn't need to know everything about threads.
And the thread manager doesn't need to know everything about GC.

At present there are two libraries, and many of the thread
related classes are actually compiled into the GC library.

There is also a potential issue spawning pthreads, duplicating
the demux event monitoring thread. This should not be necessary.
Asynchronous I/O, however, has to interlock properly with
multiple threads running separate fibre dispatchers.

With C++11 we have much more portable and advanced thread
control facilities, but to use them and clean up the system
interfaces we need more precision in the types to make the
exact couplings very explicit.