Add a safe mutex "flx_usr_mutex_t".
This mutex is a spin loop based on C++ timed_mutex
which calls the thread_control yield function in
the spin loop, ensuring requests to collection
made under a mutexi don't deadlock.

Note: using timed_mutex might be a bad choice.
Since a try_lock on a standard mutex with
a spin loop containing a wait is equivalent,
but now we only have one type of mutex .. hmm.

Note: fast_lock uses test and set on an atomic
boolean and is also a mutex.

All these functions may differ in how they yield
to the OS.