Split out dynamic link library, phase #1.
Change the namespace and link export/import macro
symbol to separate dynamic link library from
driver library, embedding library, and the core
run time library. It's still physically integrated
with the libflx (which is bad name for the core
RTL but there you have it). So it will fail on
Windows.

It should fail on Unix too but the linker visibility
control with ld is screwed.

The dynlink shape objects are not linked into the
chain of shapes. In general this doesn't matter and
isn't sustainable (except with complex messy code).