Change file modes to binary.
Note: if a foreigner opens the file, such as
the system shell, then the file, eg stdin etc
may be in text mode.

Felix NEVER opens files in text mode, not even
files called "text_files" :-) This is the
only way to get consistent cross platform behaviour.
However if the shell is using test mode for stdout, stderr,
Felix can't do anything about it.

It's up to the user to decide whether to output a CRLF
or an LF. End of line in println etc is always just LF.
C "\n" is also always just LF. So if you want Windows
line endings you have to do print "fred\r\n".