Add toolchain_msvc_cygwin.
This should allow using Visual Studio compiler
with Cygwin built Felix. Of course you cannot just do

  flx --toolchain=toolchain_msvc_cygwin hello.flx

and expect it to work: it should compile object files
but not link, because we need the MSVC++ built RTL.
So would have to select a suitable target for it to work.
The main reason for this toolchain is to allow actually
building the RTL of that target, starting with a Cygwin
Felix.