Change windows toolchain to skip all output from successful operations.
This is NOT desirable because we lose any warning
information. We will need to write specific filters
to extract such information. But CL.EXE and friends
provide NO controls to stop them spewing the names of
the files they're processing to standard output.
And we do NOT want to force pollute our display with
that crud.