Lets check the JudySL scanner.
The module directory used by preloaded plugins
is a strdict which is a JudySL array. Since this isn't
really used anywhere else than the preload dictionaries,
if it failed we'd get a crash iff we used preloading.
bootflx doesn't preload but flx does.

Note the module dictionaries are always created
on demand, so in fact ordinary code not using preloading
wont crash, even if the JudySL_scanner is totally screwed.