Fix monitor semantics.
Now, a writer thread has to wait until
the data it write to a pchannel is read
by another thread. Previously it only had
to wait until *some* data was read, not necessarily
what it wrote.

The cost of my implementation is high. The data
is wrapped up with an acknowledgment flag
which requires a heap allocation and subsequent
deallocation. Considering malloc() and free() are probably
not lock free, in fact probably use a mutex to ensure
serialisation, this may kill performance of the lock
free operations as well as by requiring a heap allocation.

But it had to be done somehow and this was the easiest way.

A lock free bag based communication channel may be useful
as well as the stricter pchannels.