Rework GUI. The technology is a major upgrade.
We use deferred drawing. Draw operations are scheduled
instead of happening immediately. A final window update
does the actual rendering. This means the visuals are
entirely state dependent, that is, independent of
control flow (the scheduling can be control flow dependent).

Although non-optimal, the whole window is redrawn every update.
This ensures everything gets drawn, and drawing is idempotent.
Optimisation will come later (and is fraught with difficulties
related to ensuring the result is the same as if the whole
window were redrawn, just faster).

Although not supported yet this also means drawing can
be made thread safe, that is, it can be done simulatenously
by any number of threads: more precisely the calculations
are can be done concurrently. The actual drawing is of course
restricted to the main thread by unfortunate platform dependent
OS/GUI designs.