Remove support for async file I/O.
Posix has an interface for aio but Linux at least
implements it with threads anyhow, which is archaic crud.
Heck, 30 yo OS had DMA based file I/O.

In any case file I/O is fast enough not to bother.
And, if you really want it you can do it with ordinary
Felix threads and, say, the thread pool anyhow.

This gets rid of the worker_fifo which has bugged me
for ages.