Fix to major bug. And several minor bugs.
A major bug has been fixed in which a function returning
a child function as a closure could be implemented
as a stack function, and thus exist as a temporary,
leaving the child closure refering to a dead frame,
as the child has dynamic lifetime, but the parent only
temporary lifetime.

Also fixed a problem where a compact linear type passed
as an argument was not decoded correctly into components,
leading to a C++ compiler error.

Some more work on effects management system.

All regression and tutorial tests are passing except
3 cases in which low inlining levels lead to service
calls left in functions, which now lead to an explicit
run time diagnostic rather than a segfault.