Slices on StringPiece. Parse capture chip. swait.
StringPiece now supports (basic) slices.
The current implementation is NOT SAFE so
do not slice off the end!

The service control (svc) operation swait
suspends the current fibre until all other
active fibres are completed. swait operations stack.
swait is a synchronous operation and does NOT take
into account fibres suspended waiting for asynchronous I/O.
If such an suspension activates before an swait its completion
will be waited on by the swait, if it comes after then, obviously
it will not be. Therefore if one does an Faio::sleep in a fibre,
and another fibre does an swait, it is likely the sleeping fibre
will reactivate after the swait. Perhaps there should be
an await, which also waits for asynchronous events to complete.
Note that logically, a program's main fibre terminates with
an await, that is, the program will not complete until all
fibres are dead, including ones waiting for asynchronous I/O.

To make the parser work as a parser, and not just a recognizer,
I have added a polymorphic leaf_capture[T] chip.
Given a recognizer which reads
a Buffer, matches something, then writes a Buffer, a leaf_capture
chip has input and output of type Buffer * T. It has two
arguments, a scan device which is the recognizer, which reads
and writes only a Buffer, and a state transformer function,
which takes an initial Buffer * T and a final Buffer,
where these Buffers mark the start and end plus one of
a lexeme, and returns a Buffer * T. The returned Buffer
component should usually be the supplied end Buffer.
The T is an arbitrary state value.

The intent is that the state value be a list of LS_expr::lsexpr,
which is a Scheme like S-expr (in list.fdoc) which has
potentially distinct labels on the Leaf and Tree nodes.
As we progress down a pipeline, we Cons all the lsexpr of
each symbol onto the list. At the end of the pipeline,
we munge the list, and make it into a Tree with a label,
and that becomes a entry in the list of the containing
grammar production, thus building an AST.

By making the whole thing polymorphic, the actual label
types can be statically determined. In an experimental
parser strings would probably be used. In any case
the LS_expr can be reduced after parsing to a fully
statically typed AST.

Well .. that's the plan anyhow ...