Refactor stuff [ci skip]
I refactor stuff so that the chip package contains
only generic chips. Strangely "tryall_list" and "pipeline_list"
are generic, even more strangely optional, match_oneormore
and match_zeroormore, are generic transducers. They're not
specific to parsing.

The recognisers package defines Buffer which is a pointer
into a string, so the various recognizers are in fact
specific to the recognition componant of parsing.

The parsing package contains components which currently
use an lsexpr_t[string,string] as the parser AST. I do not
like this. Its hard to use, requiring too much decoding.

The parser grammar is also here. This is also not good.
The pipeline and alternative combinators, as well as indirection
operator, are fully generic. However the string and regexp
matching require a Buffer, so they're specific to the
recognizer system, worse, they use the leaf_capture function
designed for lsexpr. Strangely, leaf_capture is fully generic!

However string_parser passes it the recogniser match_string,
and it passes push_entry, which is currently specific
to lsexpr_t. This all needs to be fixed!

So the TODO is to remove lsexpr_t from the system.
It should only be defined in the test cases.