Fix problem with spurious variables in GADTs.
RULE. If you write => in a union constructor, that
constructor is generalised. If a union contains
any generalised constructors, the union is a GADT.

The rule for generalised constructors is that you
must write ALL the type variables as parameters.
For non-generalised constructors, the union's type
variables automatically get added as type variables
for the constructors.

For example:
  union pair[T] =
  | PUnit of unit => pair[unit] // excludes T
  | Pair[T,U] of T * pair[U] => pair[T * pair[U]]
    // the T must be written, the union's T is not included

but
  union opt[T] = | None | Some of T

the T is included for both None and Some.