Fix uniq cstring bug hopefully.
First, add flx_strdup to strutil library, since strdup
isn't C/C++ standard.

Now, change _unsafe_cstr so it safely returns a
malloc'd copy of the input strings internal array.

Change cstr, so that hopefully it returns a GC
managed +char, this is copied AGAIN unfortunately.
That could be fixed if the GC has a "take ownership"
function, but it doesn't. It could also be fixed
if the function was part of the GC itself.
However it all should work:

(a) the input string cannot go out of scope
until after the copy is made
(b) the varray which makes the string owned by
the GC is only copying the already safely
malloc'd char array, and the freeing it.