Linear typing progress.
A number of changes to improve the detector,
plus some forced optimisations.

The ustr test case, the only one, is failing because
of low level error detecting a val containing a uniq
value. The val is actually a pointer which should be
harmless. One of the optimisations now forced may
allows the "only vars handle uniqs" to be weakened
or even lifted.

The parallel assignment routine which is used mainly
in the tail-rec optimiser, has been modified to emit
store at operations instead of assignments to projections,
which are no longer allowed. There were two cases, but
only the one I had a test for has been fixed, the other
remains with a warning diagnostic.

The system now uses the _storeat intrinsic as the default
overload for storeat procedure.

We hope to eventually remove init and assign completely
from the core system, using storeat only. Assignment
can be deduced from the form of a storeat instruction
as a special case. The code generator can use this
special case to generate C/C++ assignments instead
of storing at pointers which are, in the end, just taking
the address of the target anyhow. This may help the target
C/C++ compiler optimise its output better.