Try to fix some problems with typeof.
Seems to work some of the time, getting rid of
some spurious error messages. The method is to
wrap typeofs up so they're ignored during the first
and second phases of typedef processing and add
a third phase to expand them. The machinery
can still fail if the evaluation of an expression
to find its type depends on a typedef for which
the typeof has not yet been expanded.