Add coercions to make polymorphic variants easier to use.
We coerce variable initialisations, plain assignments,
and function returns, when the type of the variable
or function return are explicitly specified.

Assignments and initialisation ONLY allow polymorphic
variants to be coerced, however function returns use
fully general subtyping coercions, because the code
to insert the coercion in that case is in the desugaring
routine early in the compilation process before binding.