Temporaril add an rfor construct.
This construct is to test if we can fix a lot of the
pain involved with closures created in loops which
refer to data in the loop context which contains mutable
variables. When invoked the closures capture
the context which is a single frame containing the
variable holding it final value.

With the recursive implementation, the compiler creates
a separate local frame for each iteration, unless it can
determine that the self-tail call optimisation can be
applied.