Make Felix work with Ocaml 4.06.0. Add -unsafe-string switch.
Both the Felix and Fbuild build systems now compile
with -unsafe-string switch so that, primarily,
Dypgen continues to compile. It's unfortunate, this
will break Ocamls less than 4.02 since they don't
support that switch. Or Bytes for that matter.

Fixing most of Felix and OCS Scheme library seem
viable. However fixing Dypgen appears very hard
because it mixes string and bytes types left right
and centre due to the fact lexbufs are bytes
but most of the other string processing would
be the same.

Garret changed Dypgen to use Bytes but it appears
didn't test with -safe-string and the changes
do not work with 4.06 where that is the default.
A proper conversion involves looking at every
case of mixed types and trying to figure out
how to translate between them. Lexbufs are intrinsically
mutable and should have used Buffer in the first place.
Unfortunately string literals are type string and
don't mix with bytes. HOF's have to choose one or
the other. Converting lexbuf uses "to_string" is
not acceptable since that would copy the whole buffer
every access.

Basically, Ocaml team screwed up very badly.
A very stupid change.