Add chips: generator, iterate.
The name iterate is unfortunate but we
can't use iterator because iterators are
generic, that is, they're found by ordinary
overloading. If we used iterator as the chip
name it would prevent saying

   iterator x.iterator

because that would translate to

  iterator (iterator x)

and the inner iterator and outer iterator find
the same function. This is because the typeclass
iterator virtual is initially parametric but
the chip name iterator is concrete, not virtual.