Allow uniq to the thrown away in some circumstances.
It is at least safe to pass a uniq value as an argument
to a function or procedure that has a non-uniq parameter
because the function believes the parameter may be shared,
and the fact it isn't is just the "actually isn't" case
of the "maybe". If the routine is a procedure which modifies
the value we have already lost transparency. The modification
could be deliberate: var px = new "kjhg"; upcase px;.
If it happens the value is uniq, then the desired side
effect is simply lost. Maybe not intended, but perfectly safe.